+++ lib/OrdinaryDiffEqExponentialRK/test/runtests.jl
@@ -1,0 +1,3 @@
+using SafeTestsets
+
+@time @safetestset "Linear-Nonlinear Krylov Methods Tests" include("linear_nonlinear_krylov_tests.jl")
