+++ lib/OrdinaryDiffEqIMEXMultistep/src/alg_utils.jl
@@ -1,0 +1,4 @@
+issplit(alg::Union{CNAB2, CNLF2}) = true
+
+alg_order(alg::CNAB2) = 2
+alg_order(alg::CNLF2) = 2
