+++ lib/OrdinaryDiffEqLinear/src/integrator_interface.jl
@@ -1,0 +1,5 @@
+@inline function DiffEqBase.get_tmp_cache(integrator,
+        alg::LinearExponential,
+        cache::OrdinaryDiffEqMutableCache)
+    (cache.tmp,)
+end
