+++ lib/OrdinaryDiffEqLinear/test/runtests.jl
@@ -1,0 +1,3 @@
+using SafeTestsets
+
+@time @safetestset "Linear Methods Tests" include("linear_method_tests.jl")
